cdk/pyproject.toml
[project]
name = "demo-cdk"
version = "0.0.0"
dependencies = [
    "aws-cdk-lib==2.159.1",
    "boto3>=1.34.162",
    "constructs~=10.0",
]
